# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: EcoMóstoles B2B API
  description: API REST para el intercambio de residuos
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/v1/offers/{id}:
    get:
      tags:
      - Offers
      summary: Get offer by ID
      description: Retrieves the complete detail of a single material exchange offer.
        Returns 404 if the offer does not exist.
      operationId: getOfferById
      parameters:
      - name: id
        in: path
        description: Database primary key of the offer
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Offer found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferDTO"
        "404":
          description: Offer not found
        "500":
          description: Unexpected server error
    put:
      tags:
      - Offers
      summary: Update an existing offer
      description: "Replaces the editable fields of an offer. Company ownership, visit\
        \ count and publication date are preserved. Returns 404 if the offer does\
        \ not exist."
      operationId: updateOffer
      parameters:
      - name: id
        in: path
        description: Database primary key of the offer to update
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        description: "New offer values (company, visits and publicationDate are ignored)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferDTO"
        required: true
      responses:
        "200":
          description: Offer updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferDTO"
        "400":
          description: Validation failed -- check request body
        "404":
          description: Offer not found
        "500":
          description: Unexpected server error
    delete:
      tags:
      - Offers
      summary: Delete an offer
      description: Permanently removes an offer from the platform. Returns 204 NO
        CONTENT on success. Returns 404 if the offer does not exist.
      operationId: deleteOffer
      parameters:
      - name: id
        in: path
        description: Database primary key of the offer to delete
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: Offer deleted successfully
        "404":
          description: Offer not found
        "500":
          description: Unexpected server error
...
  /api/v1/offers:
    get:
      tags:
      - Offers
      summary: "List offers (paginated, optional keyword filter)"
      description: Returns a paginated list of material exchange offers. Provide `keyword`
        to filter by title or description (ACTIVE offers only). Omit `keyword` to
        retrieve all offers across all statuses.
      operationId: getAllOffers
      parameters:
      - name: keyword
        in: query
        description: Optional free-text keyword to search by title or description
          (returns only ACTIVE offers when provided)
        required: false
        schema:
          type: string
        example: plastico
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 12
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "publicationDate,DESC"
          items:
            type: string
      responses:
        "200":
          description: Page of offers returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferSummary"
        "500":
          description: Unexpected server error
    post:
      tags:
      - Offers
      summary: Create a new offer
      description: "Publishes a new material exchange offer. Returns 201 CREATED with\
        \ a Location header pointing to the new resource. Fields 'id', 'publicationDate'\
        \ and 'visits' are set server-side."
      operationId: createOffer
      requestBody:
        description: "Offer data to publish (id, publicationDate and visits are ignored)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferDTO"
        required: true
      responses:
        "201":
          description: Offer created successfully -- see Location header
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferDTO"
        "400":
          description: Validation failed -- check request body
        "500":
          description: Unexpected server error
...
  /api/v1/images/offers/{id}:
    get:
      tags:
      - Images
      summary: Get offer image
      description: Retrieves the image binary data for a specific offer.
      operationId: getOfferImage
      parameters:
      - name: id
        in: path
        description: ID of the offer
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Image returned successfully
          content:
            image/jpeg: {}
        "404":
          description: Offer or image not found
    post:
      tags:
      - Images
      summary: Upload offer image
      description: Uploads and saves a new image for a specific offer.
      operationId: uploadOfferImage
      parameters:
      - name: id
        in: path
        description: ID of the offer
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: Image file to upload
              required:
              - imageFile
      responses:
        "204":
          description: Image uploaded successfully
        "400":
          description: Invalid file or empty upload
        "404":
          description: Offer not found
...
  /api/v1/sustainability/calculate-impact:
    get:
      tags:
      - Sustainability
      summary: Calculate estimated CO2 impact
      description: Calculates the CO2 reduction in Kg based on material quantity and
        category.
      operationId: calculateImpact
      parameters:
      - name: quantity
        in: query
        description: Quantity of material
        required: true
        schema:
          type: number
          format: double
        example: 500
      - name: category
        in: query
        description: Waste category
        required: true
        schema:
          type: string
        example: METAL_WASTE
      responses:
        "200":
          description: Calculation successful
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
components:
  schemas:
    CompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commercialName:
          type: string
          minLength: 1
        contactEmail:
          type: string
          minLength: 1
        taxId:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        industrialSector:
          type: string
          minLength: 1
        description:
          type: string
          maxLength: 255
          minLength: 0
        role:
          type: string
        co2Saved:
          type: number
          format: double
        ranking:
          type: integer
          format: int32
        sector:
          type: string
        verified:
          type: boolean
        initial:
          type: string
        formattedCo2Saved:
          type: string
      required:
      - address
      - commercialName
      - contactEmail
      - description
      - industrialSector
      - phone
      - taxId
    OfferDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        wasteCategory:
          type: string
        quantity:
          type: number
          format: double
        unit:
          type: string
        price:
          type: number
          format: double
        availability:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - RESERVED
          - PAUSED
          - IN_NEGOTIATION
          - FINISHED
          - REPORTED
        publicationDate:
          type: string
          format: date-time
        visits:
          type: integer
          format: int32
        company:
          $ref: "#/components/schemas/CompanyDTO"
        closed:
          type: boolean
        category:
          type: string
        formattedPrice:
          type: string
        formattedWasteType:
          type: string
        formattedQuantity:
          type: string
        formattedPublicationDate:
          type: string
...
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commercialName:
          type: string
          minLength: 1
        taxId:
          type: string
          minLength: 1
          pattern: "^[ABCDEFGHJKLMNPQSVW][0-9]{7}[0-9A-J]$"
        contactEmail:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        industrialSector:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        logo:
          type: string
          format: byte
        roles:
          type: array
          items:
            type: string
        verified:
          type: boolean
        role:
          type: string
        admin:
          type: boolean
        client:
          type: boolean
        initial:
          type: string
      required:
      - address
      - commercialName
      - contactEmail
      - description
      - industrialSector
      - phone
      - taxId
    OfferSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        closed:
          type: boolean
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - RESERVED
          - PAUSED
          - IN_NEGOTIATION
          - FINISHED
          - REPORTED
        title:
          type: string
        quantity:
          type: number
          format: double
        unit:
          type: string
        price:
          type: number
          format: double
        company:
          $ref: "#/components/schemas/Company"
        visits:
          type: integer
          format: int32
        publicationDate:
          type: string
          format: date-time
        wasteCategory:
          type: string
        availability:
          type: string
        category:
          type: string
        formattedPrice:
          type: string
        formattedWasteType:
          type: string
        formattedQuantity:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Introduce el token JWT obtenido en /api/v1/auth/login
      in: header
      scheme: bearer
      bearerFormat: JWT